pyproject.toml
[project]
name = "zoom-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "claude-code-sdk>=0.0.19",
    "python-dotenv>=1.1.1",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
]
